New Images, AssetManager
Updated images and added tree images
Textures and fonts preloaded by AssetManager
Variables are now private where possible
Floats used for (x, y) coordinate instead of doubles
World now a public object stored in Screens.java

+++ core/src/com/tobytallis/stickygolf/Box.java
@@ -5,10 +5,10 @@ public class Box {
     int height;
     int width;
     float angle;
-    double x;
-    double y;
+    float x;
+    float y;
 
-    public Box(double x, double y, int width, int height, float angle) {
+    public Box(float x, float y, int width, int height, float angle) {
         this.x = x;
         this.y = y;
         this.width = width;
