+++ core/src/com/tobytallis/stickygolf/Button.java
@@ -6,12 +6,12 @@ public class Button {
 
     int radius;
     float angle;
-    double x;
-    double y;
+    float x;
+    float y;
     Texture texture;
     int target;
 
-    public Button(double x, double y, int radius, float angle, Texture texture, int target) {
+    public Button(float x, float y, int radius, float angle, Texture texture, int target) {
         this.x = x;
         this.y = y;
         this.radius = radius;
