+++ core/src/com/tobytallis/stickygolf/Platform.java
@@ -10,11 +10,11 @@ public class Platform {
     int width;
     // links to action (0 is no action)
     int link;
-    double x;
-    double y;
+    float x;
+    float y;
     float angle;
 
-    public Platform(double x, double y, int width, int height, float angle, boolean shadow, int[] trees, int link, boolean finalPlatform) {
+    public Platform(float x, float y, int width, int height, float angle, boolean shadow, int[] trees, int link, boolean finalPlatform) {
         this.finalPlatform = finalPlatform;
         this.x = x;
         this.y = y;
@@ -26,15 +26,15 @@ public class Platform {
         this.link = link;
     }
 
-    public Platform(double x, double y, int width, int height) {
+    public Platform(float x, float y, int width, int height) {
         this(x, y, width, height, 0, true, new int[0], 0, false);
     }
 
-    public Platform(double x, double y, int width, int height, int[] trees) {
+    public Platform(float x, float y, int width, int height, int[] trees) {
         this(x, y, width, height, 0, true, trees, 0, false);
     }
 
-    public Platform(double x, double y, int width, int height, boolean shadow) {
+    public Platform(float x, float y, int width, int height, boolean shadow) {
         this(x, y, width, height, 0, shadow, new int[0], 0, false);
     }
 }
