+++ core/src/com/tobytallis/stickygolf/Text.java
@@ -4,12 +4,12 @@ public class Text {
 
     int font;
     String text;
-    int x;
-    int y;
+    float x;
+    float y;
     MovingPlatform platformToFollow;
     Box boxToFollow;
 
-    public Text(int font, String text, int x, int y, MovingPlatform platformToFollow, Box boxToFollow) {
+    public Text(int font, String text, float x, float y, MovingPlatform platformToFollow, Box boxToFollow) {
         this.font = font;
         this.text = text;
         this.x = x;
@@ -17,15 +17,15 @@ public class Text {
         this.platformToFollow = platformToFollow;
     }
 
-    public Text(int font, String text, int x, int y, MovingPlatform platformToFollow) {
+    public Text(int font, String text, float x, float y, MovingPlatform platformToFollow) {
         this(font, text, x, y, platformToFollow, null);
     }
 
-    public Text(int font, String text, int x, int y, Box boxToFollow) {
+    public Text(int font, String text, float x, float y, Box boxToFollow) {
         this(font, text, x, y, null, boxToFollow);
     }
 
-    public Text(int font, String text, int x, int y) {
+    public Text(int font, String text, float x, float y) {
         this(font, text, x, y, null, null);
     }
 }
